requirements.txt
# Insider selling tracker - S&P 500
requests>=2.28.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
